Update UDPData_575.txt
This sentence really doesn’t make any sense. It says it hurts his general usage while giving examples of enemies that he targets. Elder sloth and calamary both have 500 range, so it isn’t a matter of relatively low range for a backliner because there is not a single backliner that can outrange them by definition of backliner being between 400 and 500. So, I changed it to be enemies he does not target and enemies that higher ranged backliners can actually outrange.